refactor main monolith
